refactor: improve CustomController and custom_grouped
