+++ lib/src/widgets/custom_grouped_checkbox.dart
@@ -1,3 +1,4 @@
+import 'package:checkbox_grouped/src/common/base_grouped_widget.dart';
 import 'package:flutter/material.dart';
 
 import 'package:checkbox_grouped/src/common/custom_state_group.dart';
@@ -17,42 +18,38 @@ import 'package:checkbox_grouped/src/controller/custom_group_controller.dart';
 ///
 /// [itemExtent] : same as [itemExtent] of [ListView]
 ///
-class CustomGroupedCheckbox<T> extends StatefulWidget {
-  final CustomGroupController controller;
+class CustomGroupedCheckbox<T> extends BaeCustomGrouped {
   final Widget? groupTitle;
   final CustomItemIndexedWidgetBuilder itemBuilder;
   final double? itemExtent;
-  final List<T> values;
   final bool _isGrid;
   final bool isScroll;
   final SliverGridDelegate? gridDelegate;
 
   CustomGroupedCheckbox({
-    Key? key,
-    required this.controller,
+    super.key,
+    required super.controller,
     this.groupTitle,
     required this.itemBuilder,
-    required this.values,
+    required super.values,
     this.isScroll = false,
     this.itemExtent,
   })  : this._isGrid = false,
-        this.gridDelegate = null,
-        super(key: key);
+        this.gridDelegate = null;
 
   CustomGroupedCheckbox.grid({
-    Key? key,
-    required this.controller,
+    super.key,
+    required super.controller,
     this.groupTitle,
     SliverGridDelegate gridDelegate = const SliverGridDelegateWithFixedCrossAxisCount(
       crossAxisCount: 3,
     ),
     this.isScroll = false,
     required this.itemBuilder,
-    required this.values,
+    required super.values,
   })  : this._isGrid = true,
         this.gridDelegate = gridDelegate,
-        this.itemExtent = 0.0,
-        super(key: key);
+        this.itemExtent = 0.0;
 
   @override
   CustomGroupedCheckboxState createState() => CustomGroupedCheckboxState();
@@ -77,60 +74,7 @@ class CustomGroupedCheckbox<T> extends StatefulWidget {
 class CustomGroupedCheckboxState<T> extends CustomStateGroup<T?, CustomGroupedCheckbox> {
   SliverChildBuilderDelegate? childrenDelegate;
 
-  @override
-  void initState() {
-    super.initState();
-    itemsSelections = ValueNotifier([]);
-    items = [];
-    widget.values.forEach((v) {
-      items.add(
-        ValueNotifier(
-          CustomItem(
-            data: v,
-            checked: widget.controller.initSelectedItem.contains(v),
-            isDisabled: false,
-          ),
-        ),
-      );
-    });
-    widget.controller.init(this);
-    if (!widget.controller.isMultipleSelection) {
-      if (widget.controller.initSelectedItem.isNotEmpty &&
-          widget.controller.initSelectedItem.first != null)
-        itemSelected.value = widget.controller.initSelectedItem.first;
-    } else {
-      if (widget.controller.initSelectedItem.isNotEmpty) {
-        itemsSelections.value = List.castFrom(widget.controller.initSelectedItem);
-      }
-    }
-  }
 
-  @override
-  void didUpdateWidget(covariant CustomGroupedCheckbox oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    if (oldWidget.controller != widget.controller) {
-      itemsSelections = ValueNotifier([]);
-      items = [];
-      widget.values.forEach((v) {
-        items.add(
-          ValueNotifier(
-            CustomItem(
-              data: v,
-              checked: widget.controller.initSelectedItem.contains(v),
-              isDisabled: false,
-            ),
-          ),
-        );
-      });
-      widget.controller.init(this);
-
-      if (!widget.controller.isMultipleSelection) {
-        itemSelected.value = widget.controller.initSelectedItem.first;
-      } else {
-        itemsSelections.value = List.castFrom(widget.controller.initSelectedItem);
-      }
-    }
-  }
 
   @override
   void dispose() {
